fix(rbac): avoid unnecessary query to check 'relations' table Refactoring ancestor search memo class. Small performance fix: avoid unnecessary query to check 'relations' table.
